Split out Artifact associated types. Don't produce a container by default - use associated for 'metadata'.
