+++ source/slang/slang-ir-metadata.cpp
@@ -4,7 +4,7 @@
 #include "slang-ir.h"
 #include "slang-ir-insts.h"
 
-#include "../compiler-core/slang-artifact-representation-impl.h"
+#include "../compiler-core/slang-artifact-associated-impl.h"
 
 namespace Slang
 {
@@ -69,11 +69,10 @@ void collectMetadata(const IRModule* irModule, PostEmitMetadata& outMetadata)
                 auto registerIndex = offsetAttr->getOffset();
                 auto size = sizeAttr->getSize();
                 auto count = size.isFinite() ? size.getFiniteValue() : 0;
-                _insertBinding(outMetadata.usedBindings, kind, spaceIndex, registerIndex, count);
+                _insertBinding(outMetadata.m_usedBindings, kind, spaceIndex, registerIndex, count);
             }
         }
     }
 }
-
 
 }
